Phase 1 (1f): harden settlement per adversarial money-path audit
An adversarial workflow audited the settlement + ledger money path (it verified
atomicity and the sum-balance invariant hold, and refuted several false positives).
Fixed the confirmed issues:

- Landing bonus/penalty now computed in decimal with away-from-zero rounding so it
  obeys the one cents convention (was double + banker's rounding: -15% of $2153.50
  gave -$323.02 instead of -$323.03; a sub-dollar greaser rounded to $0).
- Missing pilot now throws before any money is staged (was: settled silently with no
  XP recipient).
- Store-level double-settle guard: unique index on Flight.JobAssignmentId.
- Idempotent settlement postings (per-line DedupeKey settle:{id}:base/:landing).
- Tightened the aircraft-title fallback so a too-short alias cannot mis-bind.
- LedgerService.ReconcileAsync: self-heal the cached balance from the ledger (the
  recovery path the audit asked for). Full optimistic-concurrency token is deferred
  to the Phase-4 shared-world work.
- Core tests 60 -> 63.

+++ src/Callsign.Core/Economy/EconomyConfig.cs
@@ -36,16 +36,17 @@ public sealed record EconomyConfig
     // --- Settlement: landing quality (as a fraction of base reward) and payload bonus ---
     public double PayloadMatchXpBonusPct { get; init; } = 0.5;
 
-    /// <summary>Reward multiplier delta from the touchdown rate: a greaser earns a bonus, a slam a penalty.</summary>
-    public double LandingModifierPct(double touchdownFpm)
+    /// <summary>Reward multiplier delta from the touchdown rate: a greaser earns a bonus, a slam a
+    /// penalty. Decimal so the caller can round money away-from-zero, the one cents convention.</summary>
+    public decimal LandingModifierPct(double touchdownFpm)
     {
         double m = Math.Abs(touchdownFpm);
-        if (m <= 100) return 0.10;   // greaser
-        if (m <= 200) return 0.05;   // good
-        if (m <= 400) return 0.00;   // acceptable
-        if (m <= 600) return -0.05;  // firm
-        if (m <= 1000) return -0.15; // hard
-        return -0.30;                // slammed
+        if (m <= 100) return 0.10m;   // greaser
+        if (m <= 200) return 0.05m;   // good
+        if (m <= 400) return 0.00m;   // acceptable
+        if (m <= 600) return -0.05m;  // firm
+        if (m <= 1000) return -0.15m; // hard
+        return -0.30m;                // slammed
     }
 
     public static EconomyConfig Default { get; } = new();
